Renaming loaders
